Create base SQLDatabase struct for SQLite3 and PostgreSQL structs.
